+++ src/main/java/com/qa/ims/persistence/dao/ItemDAO.java
@@ -23,7 +23,7 @@ public class ItemDAO implements Dao<Item> {
 	public Item modelFromResultSet(ResultSet resultSet) throws SQLException {
 		Long itemId = resultSet.getLong("item_id");
 		String itemName = resultSet.getString("item_name");
-		Double price = resultSet.getDouble("price"); 
+		Long price = resultSet.getLong("price"); 
 		return new Item(itemId, itemName, price);
 	}
